+++ src/Arborescence.Models/GraphPolicies/MutableIndexedIncidenceGraphPolicy.cs
@@ -1,3 +1,5 @@
+#if NETSTANDARD2_1 || NETCOREAPP2_0 || NETCOREAPP2_1
+
 namespace Arborescence.Models
 {
     using System;
@@ -10,7 +12,7 @@ namespace Arborescence.Models
     public readonly struct MutableIndexedIncidenceGraphPolicy :
         ITailPolicy<MutableIndexedIncidenceGraph, int, int>,
         IHeadPolicy<MutableIndexedIncidenceGraph, int, int>,
-        IOutEdgesPolicy<MutableIndexedIncidenceGraph, int, ArrayPrefixEnumerator<int>>
+        IOutEdgesPolicy<MutableIndexedIncidenceGraph, int, ArraySegment<int>.Enumerator>
     {
         /// <inheritdoc/>
         public bool TryGetTail(MutableIndexedIncidenceGraph graph, int edge, out int tail)
@@ -31,10 +33,12 @@ namespace Arborescence.Models
         }
 
         /// <inheritdoc/>
-        public ArrayPrefixEnumerator<int> EnumerateOutEdges(MutableIndexedIncidenceGraph graph, int vertex)
+        public ArraySegment<int>.Enumerator EnumerateOutEdges(MutableIndexedIncidenceGraph graph, int vertex)
         {
-            return graph is null ? ArrayPrefixEnumerator<int>.Empty : graph.EnumerateOutEdges(vertex);
+            return graph is null ? ArraySegment<int>.Empty.GetEnumerator() : graph.EnumerateOutEdges(vertex);
         }
     }
 #pragma warning restore CA1815 // Override equals and operator equals on value types
 }
+
+#endif