+++ src/Arborescence.Models/Indexed/MutableIndexedIncidenceGraph.cs
@@ -1,11 +1,13 @@
-﻿namespace Arborescence.Models
+﻿#if NETSTANDARD2_1 || NETCOREAPP2_0 || NETCOREAPP2_1
+
+namespace Arborescence.Models
 {
     using System;
     using System.Diagnostics;
 
     /// <inheritdoc cref="Arborescence.IIncidenceGraph{TVertex, TEdge, TEdges}"/>
     public sealed class MutableIndexedIncidenceGraph :
-        IIncidenceGraph<int, int, ArrayPrefixEnumerator<int>>,
+        IIncidenceGraph<int, int, ArraySegment<int>.Enumerator>,
         IGraphBuilder<IndexedIncidenceGraph, int, int>,
         IDisposable
     {
@@ -154,13 +156,16 @@
         }
 
         /// <inheritdoc/>
-        public ArrayPrefixEnumerator<int> EnumerateOutEdges(int vertex)
+        public ArraySegment<int>.Enumerator EnumerateOutEdges(int vertex)
         {
             if (unchecked((uint)vertex >= (uint)_outEdgesByVertex.Count))
-                return ArrayPrefixEnumerator<int>.Empty;
+                return ArraySegment<int>.Empty.GetEnumerator();
 
             ArrayPrefix<int> outEdges = _outEdgesByVertex[vertex];
-            return new ArrayPrefixEnumerator<int>(outEdges.Array ?? Array.Empty<int>(), outEdges.Count);
+            if (outEdges.Array is null)
+                return ArraySegment<int>.Empty.GetEnumerator();
+
+            return new ArraySegment<int>(outEdges.Array, 0, outEdges.Count).GetEnumerator();
         }
 
         /// <summary>
@@ -174,3 +179,5 @@
         }
     }
 }
+
+#endif