+++ tests/Arborescence.Models.Tests/Simple/SimpleIncidenceGraphFromMutableTest.cs
@@ -5,7 +5,7 @@
     using Models;
     using Xunit;
     using Graph = Models.SimpleIncidenceGraph;
-    using EdgeEnumerator = ArraySegmentEnumerator<Endpoints>;
+    using EdgeEnumerator = System.ArraySegment<Endpoints>.Enumerator;
 
     public sealed class SimpleIncidenceGraphFromMutableTest
     {