+++ tests/Arborescence.Models.Tests/UndirectedIndexed/UndirectedIndexedIncidenceGraphFromMutableTest.cs
@@ -5,7 +5,7 @@ namespace Arborescence
     using Models;
     using Xunit;
     using Graph = Models.UndirectedIndexedIncidenceGraph;
-    using EdgeEnumerator = ArraySegmentEnumerator<int>;
+    using EdgeEnumerator = System.ArraySegment<int>.Enumerator;
 
     public sealed class UndirectedIndexedIncidenceGraphFromMutableTest
     {